refactor: MVC 패턴에 dto, service까지 추가
