+++ src/main/java/controller/RacingGameController.java
@@ -1,0 +1,5 @@
+package controller;
+
+public class RacingGameController {
+
+}
