+++ src/main/java/controller/dto/CarNameRequest.java
@@ -1,0 +1,5 @@
+package controller.dto;
+
+public class CarNameRequest {
+
+}
