+++ src/main/java/domain/CarName.java
@@ -1,4 +1,4 @@
-package type;
+package domain;
 
 import exception.ExceptionHandler;
