+++ src/test/java/domain/CarNameTest.java
@@ -1,4 +1,4 @@
-package type;
+package domain;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
