+++ src/test/java/domain/CarTest.java
@@ -1,4 +1,4 @@
-package type;
+package domain;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
